Reaction when sushi hits mouth, audio and particle effects are now possible on impact.

+++ Assets/Scripts/CC_Level.cs
@@ -12,7 +12,7 @@ public class CC_Level : MonoBehaviour {
 
 	public bool DebugMode = true;
 	
-	public int playerScore = 0;
+	public long playerScore = 0;
 	public int pointsToAdd = 0;
 	
 	//points needed to beat the level
